+++ btc.com/src/adapter.ts
@@ -1,0 +1,40 @@
+import { Requester, Validator, AdapterError } from '@chainlink/external-adapter'
+import { ExecuteWithConfig, ExecuteFactory } from '@chainlink/types'
+import { DEFAULT_ENDPOINT } from './config'
+import { balance } from './endpoint'
+
+const inputParams = {
+  endpoint: false,
+}
+
+// Export function to integrate with Chainlink node
+export const execute: ExecuteWithConfig = async (request, config) => {
+  const validator = new Validator(request, inputParams)
+  if (validator.error) throw validator.error
+
+  Requester.logConfig(config)
+
+  const jobRunID = validator.validated.id
+  const endpoint = validator.validated.data.endpoint || DEFAULT_ENDPOINT
+
+  let response
+  switch (endpoint) {
+    case balance.Name: {
+      response = await balance.execute(config, request)
+      break
+    }
+    default: {
+      throw new AdapterError({
+        jobRunID,
+        message: `Endpoint ${endpoint} not supported.`,
+        statusCode: 400,
+      })
+    }
+  }
+
+  return Requester.success(jobRunID, response)
+}
+
+export const makeExecute: ExecuteFactory = (config) => {
+  return async (request) => execute(request, config || Requester.getDefaultConfig())
+}