+++ btc.com/src/config.ts
@@ -1,0 +1,6 @@
+export const ENDPOINT_MAIN = 'https://chain.api.btc.com'
+
+export const DEFAULT_DATA_PATH = 'addresses'
+export const DEFAULT_ENDPOINT = 'balance'
+
+export const getBaseURL = (): string => ENDPOINT_MAIN