+++ btc.com/test/balance.test.ts
@@ -1,0 +1,4 @@
+import { makeExecute } from '../src/adapter'
+import { shouldBehaveLikeBalanceAdapter } from '@chainlink/adapter-test-helpers'
+
+shouldBehaveLikeBalanceAdapter(makeExecute(), ['bitcoinmainnet'])